ToneGenBase in lib-builtin-effects
